requirements.txt
streamlit
pandas
numpy
snowflake-connector-python
google-generativeai 
streamlit-extras
dotenv